% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_type.R
\name{change_type}
\alias{change_type}
\title{Change variable type in a nested dataframe}
\usage{
change_type(col, name_col)
}
\arguments{
\item{col}{The columns to change of the nested dataframe}

\item{name_col}{A string of the name of the column to change}
}
\value{
The columns with a new type
}
\description{
Change variable type in a nested dataframe
}
\examples{
change_type(data[i],"time")
}
