% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_type.R
\name{change_type_df}
\alias{change_type_df}
\title{Change the type of all columns in the nested dataframe}
\usage{
change_type_df(df)
}
\arguments{
\item{df}{Nested dataframe to change variables types}
}
\value{
Nested dataframe with types variables changed
}
\description{
Change the type of all columns in the nested dataframe
}
\examples{
change_type_df(data)
}
