% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_type.R
\name{change_type_user}
\alias{change_type_user}
\title{Change type according the input of the user}
\usage{
change_type_user(type, col)
}
\arguments{
\item{type}{The type desired by the user}

\item{col}{The columns to apply the new type}
}
\value{
The same columns with a new type
}
\description{
Change type according the input of the user
}
\examples{
change_type_ser("int",data[i])
change_type_ser("date",data[i])
}
