% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funnel_plot.R
\name{check_missing}
\alias{check_missing}
\title{Check is NA and replace by 0}
\usage{
check_missing(colonne, data, data2)
}
\arguments{
\item{colonne}{A column}

\item{data}{A dataframe}

\item{data2}{An other dataframe}
}
\value{
A dataframe
}
\description{
Check is NA and replace by 0
}
\examples{
check_missing(myColumns,myDF,myDF2)
}
