% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onUnload}
\alias{.onUnload}
\title{Remove path to the Shiny App}
\usage{
.onUnload(libname, pkgname)
}
\arguments{
\item{libname}{Name of the library}

\item{pkgname}{Name of the package}
}
\value{
Nothing
}
\description{
Remove path to the Shiny App
}
\examples{
.onUnLoad(libname, "FootballeuR")
}
