% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lineup.R
\name{formation_coord}
\alias{formation_coord}
\title{Get the coordinates of players according the formation}
\usage{
formation_coord(formation)
}
\arguments{
\item{formation}{A string}
}
\value{
A dataframe with 2 columns x and y
}
\description{
Get the coordinates of players according the formation
}
\examples{
formation_coord("442")
}
