% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lineup.R
\name{get_lineups}
\alias{get_lineups}
\title{Display lineup on a soccer pitch}
\usage{
get_lineups(df)
}
\arguments{
\item{df}{A dataframe}
}
\value{
The lineups
}
\description{
Display lineup on a soccer pitch
}
\examples{
get_lineups(df)
}
