% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_score.R
\name{get_score}
\alias{get_score}
\title{Get a dataframe of all scoreurs of a single game}
\usage{
get_score(df)
}
\arguments{
\item{df}{A dataframe}
}
\value{
A dataframe
}
\description{
Get a dataframe of all scoreurs of a single game
}
\examples{
get_score(df)
}
