% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_player.R
\name{heatmap_player}
\alias{heatmap_player}
\title{Heatmap on soccer pitch to describe activity by player}
\usage{
heatmap_player(df, name_player)
}
\arguments{
\item{df}{A Dataframe}

\item{name_player}{The name of the player among all present in the game}
}
\value{
An heatmap on a soccer pitch
}
\description{
Heatmap on soccer pitch to describe activity by player
}
\examples{
heatmap_player(df,"Lionel Andrés Messi Cuccittini")
}
