% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_packages.R
\name{load_packages}
\alias{load_packages}
\title{Load required packages for our package}
\usage{
load_packages(packages)
}
\arguments{
\item{packages}{A vector of string}
}
\value{
Nothing
}
\description{
Load required packages for our package
}
\examples{
load_packages(c("dplyr", "ggplot2"))
}
