% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val_manquantes.R
\name{missing_data}
\alias{missing_data}
\title{Count missing data by columns}
\usage{
missing_data(df)
}
\arguments{
\item{df}{A dataframe}
}
\value{
Nothing
}
\description{
Count missing data by columns
}
\examples{
missing_data(myDF)
}
