% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val_manquantes.R
\name{missing_data_list}
\alias{missing_data_list}
\title{Count all missing data in a list}
\usage{
missing_data_list(l)
}
\arguments{
\item{l}{A list}
}
\value{
A frequency
}
\description{
Count all missing data in a list
}
\examples{
missing_data_list(myList)
}
