% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lineup.R
\name{position_player}
\alias{position_player}
\title{Get the data lineups and coordinates}
\usage{
position_player(formation, data_lineup, team_side)
}
\arguments{
\item{formation}{A string}

\item{data_lineup}{Data lineup}

\item{team_side}{A value (1 or 2)}
}
\value{
Dataframe with lineups and coordinates
}
\description{
Get the data lineups and coordinates
}
\examples{
position_player("442",data_lineup,1)
}
