% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radar_plot.R
\name{radar_plot}
\alias{radar_plot}
\title{Radar plot on player contribution}
\usage{
radar_plot(df)
}
\arguments{
\item{df}{A dataframe}
}
\value{
A radar plot
}
\description{
Radar plot on player contribution
}
\examples{
radar_plot(myDF)
}
