% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_type.R
\name{range_list}
\alias{range_list}
\title{Function to range a list}
\usage{
range_list(l)
}
\arguments{
\item{l}{A list}
}
\value{
The same list with type changed
}
\description{
Function to range a list
}
\examples{
range_list(l)
}
