% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.R
\name{shot_pitch}
\alias{shot_pitch}
\title{Function to display shots of a game}
\usage{
shot_pitch(tble, time)
}
\arguments{
\item{tble}{The dataframe}

\item{time}{A vector with begin and end time}
}
\value{
A soccer pitch with points represen by shots
}
\description{
Function to display shots of a game
}
\examples{
shot_pitch(df,c(0,90))
shot_pitch(df,c(0,45))
shot_pitch(df,c(45,90))
}
