% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supColonne.R
\name{supCol_by_names}
\alias{supCol_by_names}
\title{Delete columns of dataframe according their names}
\usage{
supCol_by_names(df, col_names)
}
\arguments{
\item{df}{The dataframe to update.}

\item{col_names}{A vector of string which contains the names of columns to drop}
}
\value{
A dataframe.
}
\description{
Delete columns of dataframe according their names
}
\examples{
test <- jsonlite::fromJSON("../data-raw/15946.json")
remove_columns_by_name(test,c("id","index","related_events"))
}
