% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateNameFiles.R
\name{updateNameFiles}
\alias{updateNameFiles}
\title{Update names of few dataframes}
\usage{
updateNameFiles(df, col)
}
\arguments{
\item{DFs}{A dataframe}

\item{colonne}{A columns of the dataframe}
}
\value{
A name
}
\description{
Update names of few dataframes
}
\examples{
updateNameFiles(df1,"team")
}
