% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lineup.R
\name{get_lineups}
\alias{get_lineups}
\title{Display lineup on a soccer pitch}
\usage{
get_lineups(df, maillot_dom, maillot_ext)
}
\arguments{
\item{df}{A dataframe}

\item{maillot_dom}{The path to the image of the jersey}

\item{maillot_ext}{The path to the image of the jersey of the visitors}
}
\value{
The lineups
}
\description{
Display lineup on a soccer pitch
}
\examples{
get_lineups(df,maillot_dom,maillot_ext)
}
