% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateNameFiles.R
\name{updateNameFiles}
\alias{updateNameFiles}
\title{Update names of few dataframes}
\usage{
updateNameFiles(DFs, col)
}
\arguments{
\item{DFs}{A list of dataframes}

\item{colonne}{A columns of the dataframe}
}
\value{
A list of the names of the dataframes
}
\description{
Update names of few dataframes
}
\examples{
updateNameFiles(`../data`,"team")
}
